backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
sqlalchemy>=2.0.36
pydantic>=2.9.2
pydantic-settings>=2.6.1
python-multipart>=0.0.12
chromadb>=0.5.20
langchain>=0.3.7
langchain-google-genai>=2.0.4
langchain-community>=0.3.7
langchain-chroma>=0.1.4
google-generativeai>=0.8.3
pypdf>=5.1.0
python-docx>=1.1.2
markdown>=3.7
python-dotenv>=1.0.1
aiofiles>=24.1.0
